requirements.txt
docutils==0.21.2
faiss-cpu==1.13.2
fastapi
ffmpy==1.0.0
huggingface-hub==0.36.0
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-supervisor==0.0.31
mongoengine==0.29.1
numpy==2.0.2
openai==2.12.0
pandas==2.2.2
pydantic==2.12.3
pydantic-settings==2.12.0
pydantic_core==2.41.4
pymongo==4.15.5
python-dotenv==1.2.1
sentence-transformers==5.2.0
uvicorn
watchdog==6.0.0
pypdf==6.5.0
rank_bm25==0.2.2
gtts==2.5.4
openai-whisper==20250625